refactor(icons): extract square-icon handling into apis.shared.images
Managed models are about to attach icons the same way Agents do, and the
part that must not be written twice is the validation: format sniffing,
the square/size limits, and the always-on re-encode that strips EXIF.
A second copy would drift, and the copy that drifted would be the one
that published somebody's GPS coordinates.

Moves normalize_icon, the encode ladder, content_digest and a generic
IconStore into apis/shared/images/icons.py. assistants/icons.py keeps
what is actually about Agents — the key layout and the serve URL — and
re-exports the rest, so no caller changes.

Pure move: no behavior change. The two agent-icon tests that reached for
the private _encode_within_limit now point at its new home.
